safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/network/network-clear-cache-expected.txt

CONSOLE MESSAGE: line 13: First XHR loaded.
CONSOLE MESSAGE: line 28: Done.
Tests clearing browser cache from inspector.

http://127.0.0.1:8000/inspector/network/resources/random-cached.php
http://127.0.0.1:8000/inspector/network/resources/random-cached.php
http://127.0.0.1:8000/inspector/network/resources/random-cached.php
